+++ tests/kedro_neptune/test_standard.py
@@ -13,6 +13,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+from datetime import datetime
+
 from tests.kedro_neptune.utils.kedro_utils import run_pipeline
 from tests.kedro_neptune.utils.run_utils import assert_structure
 
@@ -25,3 +27,12 @@ def test_run():
 def test_run_with_params():
     run_pipeline(project="planets", session_params={"extra_params": {"travel_speed": 40000}})
     assert_structure(travel_speed=40000)
+
+
+def test_run_with_custom_run_id():
+    custom_run_id = str(datetime.now())
+    run_pipeline(
+        project="planets",
+        custom_run_id=custom_run_id,
+    )
+    assert_structure(custom_run_id=custom_run_id)