+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.5.0
+
+### Features
+- Added support for the `NEPTUNE_CUSTOM_RUN_ID` environment variable ([#81](https://github.com/neptune-ai/kedro-neptune/pull/81))
+
 ## 0.4.0
 
 ### Features